Working BatchInsert
git-svn-id: https://svn.princeton.com/svn/pci_share/Dataload@320 f6c9f1ed-0ca6-2e48-b550-0345cde077c8

+++ lib/dataload/table_manager.rb
@@ -1,6 +1,7 @@
 class TableManager
   include TableModule
   def delete_rows!
+    return unless ar_cls.table_exists?
     ar_cls.connection.execute("DELETE from #{table_name}")
     puts "Deleted rows from #{table_name}"
   end
